+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Fixed
 - Fixed identifiers conflicting with Python keywords ([issue#6]).
 - Fixed broken version comparison in `_winrt.pyi`.
+- Fixed `@typing.overload` used on methods that are not overloaded.
 
 [issue#6]: https://github.com/pywinrt/pywinrt/issues/6
 [issue#7]: https://github.com/pywinrt/pywinrt/issues/7